--- a/tsdb/head_wal.go
+++ b/tsdb/head_wal.go
@@ -500,12 +500,8 @@ func (h *Head) resetSeriesWithMMappedChunks(mSeries *memSeries, mmc, oooMmc []*m
 	h.metrics.chunks.Add(float64(len(mmc) + len(oooMmc) - len(mSeries.mmappedChunks)))
 	mSeries.mmappedChunks = mmc
 	mSeries.oooMmappedChunks = oooMmc
-	// Cache the last mmapped chunk time, so we can skip calling append() for samples it will reject.
-	if len(mmc) == 0 {
-		mSeries.mmMaxTime = math.MinInt64
-	} else {
-		mSeries.mmMaxTime = mmc[len(mmc)-1].maxTime
-		h.updateMinMaxTime(mmc[0].minTime, mSeries.mmMaxTime)
+	if len(mmc) > 0 {
+		h.updateMinMaxTime(mmc[0].minTime, mmc[len(mmc)-1].maxTime)
 	}
 	if len(oooMmc) != 0 {
 		// Mint and maxt can be in any chunk, they are not sorted.
@@ -602,10 +598,12 @@ func (wp *walSubsetProcessor) processWALSamples(h *Head, mmappedChunks, oooMmapp
 				unknownRefs++
 				continue
 			}
-			if s.T <= ms.mmMaxTime {
-				continue
+			mmMaxTime := int64(math.MinInt64)
+			mmLen := len(ms.mmappedChunks)
+			if mmLen > 0 {
+				mmMaxTime = ms.mmappedChunks[mmLen-1].maxTime
 			}
-			if s.T <= ms.mmMaxTime {
+			if s.T <= mmMaxTime {
 				continue
 			}
 			if _, chunkCreated := ms.append(s.T, s.V, 0, h.chunkDiskMapper, chunkRange); chunkCreated {
@@ -633,7 +631,12 @@ func (wp *walSubsetProcessor) processWALSamples(h *Head, mmappedChunks, oooMmapp
 				unknownHistogramRefs++
 				continue
 			}
-			if s.t <= ms.mmMaxTime {
+			mmMaxTime := int64(math.MinInt64)
+			mmLen := len(ms.mmappedChunks)
+			if mmLen > 0 {
+				mmMaxTime = ms.mmappedChunks[mmLen-1].maxTime
+			}
+			if s.t <= mmMaxTime {
 				continue
 			}
 			var chunkCreated bool