@bboreham 2022-12-28
I wonder if the benchmarks are creating data which triggers the situation where it helps?

When I made this change I was using a real WAL to check the effect.