@codesome 2022-12-28
I can try it with a real-world WAL as well.