@codesome 2023-01-05
I tried WAL replay on 8GB of WAL, and this PR is 0.8-3% slower. There were no mmap chunks on the disk, so the WAL replay also creates them in the process, so the slowdown could be the disk write or this change. Either way, I don't think that is a big difference to have the variable around.