@bboreham 2023-01-06
> There were no mmap chunks on the disk

Is that a valid test?  With 15 second scrape, mmap data is ~30 minutes old while the WAL goes back ~3 hours, so most of it gets skipped.  If you tested with no mmap files you wouldn't see this.